Move tests associated with client.eval_xslt() into their own file
